将辅助方法的返回值修改为uint32_t，并对数组定义进行常数折叠优化

+++ frontend/antlr4/Antlr4CSTVisitor.h
@@ -332,5 +332,5 @@ private:
     /// @param node ast节点
     /// @return std::optional<uint64_t>
     ///
-    std::optional<uint64_t> calculate_const_dim_size(ast_node * node);
+    std::optional<uint32_t> calculate_const_dim_size(ast_node * node);
 };
